Extract suggest_draft_pick function to separate module
- Moved suggest_draft_pick and all helper functions to src/tools/draft_suggestions.py
- Updated imports in server.py and all test files
- Fixed test mocks to use new import path
- All functionality preserved with clean separation of concerns
