+++ tests/services/test_roster_analysis_sheets_format.py
@@ -2,7 +2,7 @@ from unittest.mock import patch
 
 import pytest
 
-from src.tools.mcp_tools import suggest_draft_pick
+from src.tools.draft_suggestions import suggest_draft_pick
 
 
 class TestRosterAnalysisSheetsFormat:
@@ -78,7 +78,7 @@ class TestRosterAnalysisSheetsFormat:
             },
         }
 
-        with patch("src.tools.mcp_tools.get_player_rankings") as mock_rankings:
+        with patch("src.tools.player_rankings.get_player_rankings") as mock_rankings:
             mock_rankings.return_value = mock_rankings_result
 
             result = await suggest_draft_pick(mock_draft_state, owner_name="Test Owner")
@@ -159,7 +159,7 @@ class TestRosterAnalysisSheetsFormat:
             },
         }
 
-        with patch("src.tools.mcp_tools.get_player_rankings") as mock_rankings:
+        with patch("src.tools.player_rankings.get_player_rankings") as mock_rankings:
             mock_rankings.return_value = mock_rankings_result
 
             result = await suggest_draft_pick(mock_draft_state, owner_name="Test Owner")
