+++ tests/test_drafted_player_composite_keys.py
@@ -14,7 +14,7 @@ import pytest
 # Add src to path
 sys.path.append(str(Path(__file__).parent.parent / "src"))
 
-from tools.mcp_tools import analyze_available_players
+from src.tools import analyze_available_players
 
 
 class TestDraftedPlayerCompositeKeys:
@@ -72,7 +72,7 @@ class TestDraftedPlayerCompositeKeys:
         }
 
         with patch(
-            "tools.mcp_tools.get_player_rankings", new_callable=AsyncMock
+            "src.tools.analyze_players.get_player_rankings", new_callable=AsyncMock
         ) as mock_rankings:
             with patch("config.USER_OWNER_NAME", "Adam"):
                 mock_rankings.return_value = mock_rankings_response
@@ -143,7 +143,7 @@ class TestDraftedPlayerCompositeKeys:
         }
 
         with patch(
-            "tools.mcp_tools.get_player_rankings", new_callable=AsyncMock
+            "src.tools.analyze_players.get_player_rankings", new_callable=AsyncMock
         ) as mock_rankings:
             with patch("config.USER_OWNER_NAME", "Adam"):
                 mock_rankings.return_value = mock_rankings_response
