+++ tests/test_owner_team_lookup.py
@@ -12,7 +12,7 @@ import pytest
 # Add src to path
 sys.path.append(str(Path(__file__).parent.parent / "src"))
 
-from tools.mcp_tools import analyze_available_players
+from src.tools import analyze_available_players
 
 
 class TestOwnerTeamLookup:
@@ -50,9 +50,9 @@ class TestOwnerTeamLookup:
         }
 
         with patch(
-            "tools.mcp_tools.get_player_rankings", new_callable=AsyncMock
+            "src.tools.analyze_players.get_player_rankings", new_callable=AsyncMock
         ) as mock_rankings:
-            with patch("config.USER_OWNER_NAME", "Adam"):
+            with patch("src.tools.analyze_players.USER_OWNER_NAME", "Adam"):
                 mock_rankings.return_value = mock_rankings_response
 
                 result = await analyze_available_players(
@@ -97,9 +97,9 @@ class TestOwnerTeamLookup:
         }
 
         with patch(
-            "tools.mcp_tools.get_player_rankings", new_callable=AsyncMock
+            "src.tools.analyze_players.get_player_rankings", new_callable=AsyncMock
         ) as mock_rankings:
-            with patch("config.USER_OWNER_NAME", "Adam"):
+            with patch("src.tools.analyze_players.USER_OWNER_NAME", "Adam"):
                 mock_rankings.return_value = mock_rankings_response
 
                 result = await analyze_available_players(
@@ -144,9 +144,9 @@ class TestOwnerTeamLookup:
         }
 
         with patch(
-            "tools.mcp_tools.get_player_rankings", new_callable=AsyncMock
+            "src.tools.analyze_players.get_player_rankings", new_callable=AsyncMock
         ) as mock_rankings:
-            with patch("config.USER_OWNER_NAME", "Adam"):
+            with patch("src.tools.analyze_players.USER_OWNER_NAME", "Adam"):
                 mock_rankings.return_value = mock_rankings_response
 
                 result = await analyze_available_players(
@@ -191,9 +191,9 @@ class TestOwnerTeamLookup:
         }
 
         with patch(
-            "tools.mcp_tools.get_player_rankings", new_callable=AsyncMock
+            "src.tools.analyze_players.get_player_rankings", new_callable=AsyncMock
         ) as mock_rankings:
-            with patch("config.USER_OWNER_NAME", "adam"):
+            with patch("src.tools.analyze_players.USER_OWNER_NAME", "adam"):
                 mock_rankings.return_value = mock_rankings_response
 
                 result = await analyze_available_players(
